feat(TMDM-12742) : Resolution Where interfaces
