+++ org.talend.mdm.core/src/com/amalto/core/storage/StorageMetadataUtils.java
@@ -14,6 +14,7 @@ import com.amalto.core.query.user.ConstantExpression;
 import com.amalto.core.query.user.DateConstant;
 import com.amalto.core.query.user.DateTimeConstant;
 import com.amalto.core.query.user.TimeConstant;
+import com.amalto.core.storage.record.CommonStorageMetaDataUtils;
 import com.amalto.core.storage.record.DataRecord;
 import com.amalto.core.storage.record.metadata.UnsupportedDataRecordMetadata;
 import org.apache.commons.collections.CollectionUtils;
