+++ org.talend.mdm.query/src/com/amalto/core/util/CommonUtilXml.java
@@ -1,0 +1,110 @@
+package com.amalto.core.util;
+
+import org.w3c.dom.Document;
+import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
+import org.xml.sax.InputSource;
+import org.xml.sax.SAXException;
+
+import javax.xml.namespace.NamespaceContext;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.TransformerException;
+import javax.xml.xpath.XPath;
+import javax.xml.xpath.XPathConstants;
+import javax.xml.xpath.XPathFactory;
+import java.io.IOException;
+import java.io.StringReader;
+import java.util.Collections;
+import java.util.Iterator;
+
+public class CommonUtilXml {
+
+    private static DocumentBuilderFactory nonValidatingDocumentBuilderFactory;
+
+    private static synchronized DocumentBuilderFactory getDocumentBuilderFactory() {
+        if (nonValidatingDocumentBuilderFactory == null) {
+            nonValidatingDocumentBuilderFactory = DocumentBuilderFactory.newInstance();
+            nonValidatingDocumentBuilderFactory.setNamespaceAware(true);
+            nonValidatingDocumentBuilderFactory.setValidating(false);
+            nonValidatingDocumentBuilderFactory.setExpandEntityReferences(false);
+        }
+        return nonValidatingDocumentBuilderFactory;
+    }
+
+    public static Document parse(String xmlString) throws ParserConfigurationException, IOException, SAXException {
+        DocumentBuilderFactory factory;
+        factory = getDocumentBuilderFactory();
+        factory.setExpandEntityReferences(false);
+        DocumentBuilder builder = factory.newDocumentBuilder();
+        SAXErrorHandler seh = new SAXErrorHandler();
+        builder.setErrorHandler(seh);
+        Document d = builder.parse(new InputSource(new StringReader(xmlString)));
+        // check if document parsed correctly against the schema
+        String errors = seh.getErrors();
+        if (errors.length() != 0) {
+            String err = "Document did not parse against schema: \n" + errors + "\n"
+                    + xmlString.substring(0, Math.min(100, xmlString.length()));
+            throw new SAXException(err);
+        }
+        return d;
+    }
+
+    public static String getFirstTextNode(Node contextNode, String xPath, Node namespaceNode) throws TransformerException {
+        String[] res = getTextNodes(contextNode, xPath, namespaceNode);
+        if (res.length == 0) {
+            return null;
+        }
+        return res[0];
+    }
+
+    public static String getFirstTextNode(Node contextNode, String xPath) throws TransformerException {
+        return getFirstTextNode(contextNode, xPath, contextNode);
+    }
+
+    public static String[] getTextNodes(Node contextNode, String xPath, final Node namespaceNode) throws TransformerException {
+        String[] results;
+        // test for hard-coded values
+        if (xPath.startsWith("\"") && xPath.endsWith("\"")) {
+            return new String[] { xPath.substring(1, xPath.length() - 1) };
+        }
+        // test for incomplete path (elements missing /text())
+        if (!xPath.matches(".*@[^/\\]]+")) { // attribute
+            if (!xPath.endsWith(")")) { // function
+                xPath += "/text()";
+            }
+        }
+        try {
+            XPath path = XPathFactory.newInstance().newXPath();
+            path.setNamespaceContext(new NamespaceContext() {
+
+                @Override
+                public String getNamespaceURI(String s) {
+                    return namespaceNode.getNamespaceURI();
+                }
+
+                @Override
+                public String getPrefix(String s) {
+                    return namespaceNode.getPrefix();
+                }
+
+                @Override
+                public Iterator getPrefixes(String s) {
+                    return Collections.singleton(namespaceNode.getPrefix()).iterator();
+                }
+            });
+            NodeList xo = (NodeList) path.evaluate(xPath, contextNode, XPathConstants.NODESET);
+            results = new String[xo.getLength()];
+            for (int i = 0; i < xo.getLength(); i++) {
+                results[i] = xo.item(i).getTextContent();
+            }
+        } catch (Exception e) {
+            String err = "Unable to get the text node(s) of " + xPath + ": " + e.getClass().getName() + ": "
+                    + e.getLocalizedMessage();
+            throw new TransformerException(err);
+        }
+        return results;
+
+    }
+}
