+++ org.talend.mdm.query/src/com/amalto/core/util/SAXErrorHandler.java
