+++ org.talend.mdm.query/src/com/amalto/xmlserver/interfaces/IWhereItem.java
