+++ org.talend.mdm.query/src/com/amalto/xmlserver/interfaces/WhereCondition.java
@@ -17,6 +17,7 @@ import java.io.Serializable;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import com.amalto.core.util.CommonUtilXml;
 import org.w3c.dom.Document;
 
 /**
@@ -95,12 +96,12 @@ public class WhereCondition implements IWhereItem, Serializable {
 
     public static WhereCondition deserialize(String xml) throws XmlServerException {
         try {
-            Document d = com.amalto.core.util.Util.parse(xml);
-            return new WhereCondition(com.amalto.core.util.Util.getFirstTextNode(d.getDocumentElement(), "./leftpath"),
-                    com.amalto.core.util.Util.getFirstTextNode(d.getDocumentElement(), "./operator"),
-                    com.amalto.core.util.Util.getFirstTextNode(d.getDocumentElement(), "./rightvalueorpath"),
-                    com.amalto.core.util.Util.getFirstTextNode(d.getDocumentElement(), "./stringpredicate"),
-                    "yes".equals(com.amalto.core.util.Util.getFirstTextNode(d.getDocumentElement(), "./spellcheck")));
+            Document d = CommonUtilXml.parse(xml);
+            return new WhereCondition(CommonUtilXml.getFirstTextNode(d.getDocumentElement(), "./leftpath"),
+                    CommonUtilXml.getFirstTextNode(d.getDocumentElement(), "./operator"),
+                    CommonUtilXml.getFirstTextNode(d.getDocumentElement(), "./rightvalueorpath"),
+                    CommonUtilXml.getFirstTextNode(d.getDocumentElement(), "./stringpredicate"),
+                    "yes".equals(CommonUtilXml.getFirstTextNode(d.getDocumentElement(), "./spellcheck")));
         } catch (Exception e) {
             throw new XmlServerException(e);
         }
