+++ org.talend.mdm.query/src/com/amalto/xmlserver/interfaces/XmlServerException.java
